+++ stairsplus/registrations.lua
@@ -1,7 +1,7 @@
 --[[
 More Blocks: registrations
 
-Copyright (c) 2011-2017 Hugo Locurcio and contributors.
+Copyright (c) 2011-2018 Hugo Locurcio and contributors.
 Licensed under the zlib license. See LICENSE.md for more information.
 --]]
 
@@ -71,6 +71,8 @@ for _, name in pairs(default_nodes) do
 	mod = "moreblocks"
 	stairsplus:register_all(mod, name, nodename, ndef)
 	minetest.register_alias_force("stairs:stair_" .. name, mod .. ":stair_" .. name)
+	minetest.register_alias_force("stairs:stair_outer_" .. name, mod .. ":stair_" .. name .. "_outer")
+	minetest.register_alias_force("stairs:stair_inner_" .. name, mod .. ":stair_" .. name .. "_inner")
 	minetest.register_alias_force("stairs:slab_"  .. name, mod .. ":slab_"  .. name)
 end
 
@@ -86,6 +88,8 @@ if minetest.get_modpath("farming") then
 		mod = "moreblocks"
 		stairsplus:register_all(mod, name, nodename, ndef)
 		minetest.register_alias_force("stairs:stair_" .. name, mod .. ":stair_" .. name)
+		minetest.register_alias_force("stairs:stair_outer_" .. name, mod .. ":stair_" .. name .. "_outer")
+		minetest.register_alias_force("stairs:stair_inner_" .. name, mod .. ":stair_" .. name .. "_inner")
 		minetest.register_alias_force("stairs:slab_"  .. name, mod .. ":slab_"  .. name)
 	end
 end
@@ -106,6 +110,36 @@ if minetest.get_modpath("wool") then
 
 		stairsplus:register_all(mod, name, nodename, ndef)
 	end
+end
+
+-- basic_materials, keeping the original other-mod-oriented names
+-- for backwards compatibility
+
+if minetest.get_modpath("basic_materials") then
+	stairsplus:register_all("technic","concrete","basic_materials:concrete_block",{
+		description = "Concrete",
+		tiles = {"basic_materials_concrete_block.png",},
+		groups = {cracky=1, level=2, concrete=1},
+		sounds = default.node_sound_stone_defaults(),
+	})
+
+	minetest.register_alias("prefab:concrete_stair","technic:stair_concrete")
+	minetest.register_alias("prefab:concrete_slab","technic:slab_concrete")
+
+	stairsplus:register_all("gloopblocks", "cement", "basic_materials:cement_block", {
+		description = "Cement",
+		tiles = {"basic_materials_cement_block.png"},
+		groups = {cracky=2, not_in_creative_inventory=1},
+		sounds = default.node_sound_stone_defaults(),
+		sunlight_propagates = true,
+	})
+
+	stairsplus:register_all("technic", "brass_block", "basic_materials:brass_block", {
+		description="Brass Block",
+		groups={cracky=1, not_in_creative_inventory=1},
+		tiles={"basic_materials_brass_block.png"},
+	})
+
 end
 
 -- Alias cuts of split_stone_tile_alt which was renamed checker_stone_tile.